refactor: register workspace dropdown menu to left menu
