+++ src/core/public/index.ts
@@ -348,7 +348,13 @@ export {
 
 export { __osdBootstrap__ } from './osd_bootstrap';
 
-export { WorkspaceStart, WorkspaceService, WorkspaceAttribute } from './workspace';
+export {
+  WorkspaceStart,
+  WorkspaceSetup,
+  WorkspaceService,
+  WorkspaceAttribute,
+  WorkspaceObservables,
+} from './workspace';
 
 export { WorkspacePermissionMode, PUBLIC_WORKSPACE, MANAGEMENT_WORKSPACE } from '../utils';
