+++ src/core/public/workspace/index.ts
@@ -2,5 +2,10 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-export { WorkspaceStart, WorkspaceService, WorkspaceSetup } from './workspaces_service';
+export {
+  WorkspaceStart,
+  WorkspaceService,
+  WorkspaceSetup,
+  WorkspaceObservables,
+} from './workspaces_service';
 export type { WorkspaceAttribute } from './workspaces_service';
